move NavEvent to internal package
