+++ navigator/runtime/src/main/java/com/freeletics/mad/navigator/NavEventNavigator.kt
@@ -3,15 +3,16 @@ package com.freeletics.mad.navigator
 import android.os.Parcelable
 import androidx.activity.OnBackPressedCallback
 import androidx.activity.result.contract.ActivityResultContract
-import com.freeletics.mad.navigator.NavEvent.ActivityResultEvent
-import com.freeletics.mad.navigator.NavEvent.BackEvent
-import com.freeletics.mad.navigator.NavEvent.BackToEvent
-import com.freeletics.mad.navigator.NavEvent.NavigateToActivityEvent
-import com.freeletics.mad.navigator.NavEvent.NavigateToEvent
-import com.freeletics.mad.navigator.NavEvent.UpEvent
+import com.freeletics.mad.navigator.internal.NavEvent.ActivityResultEvent
+import com.freeletics.mad.navigator.internal.NavEvent.BackEvent
+import com.freeletics.mad.navigator.internal.NavEvent.BackToEvent
+import com.freeletics.mad.navigator.internal.NavEvent.NavigateToActivityEvent
+import com.freeletics.mad.navigator.internal.NavEvent.NavigateToEvent
+import com.freeletics.mad.navigator.internal.NavEvent.UpEvent
 import com.freeletics.mad.navigator.internal.DelegatingOnBackPressedCallback
 import com.freeletics.mad.navigator.internal.DestinationId
 import com.freeletics.mad.navigator.internal.InternalNavigatorApi
+import com.freeletics.mad.navigator.internal.NavEvent
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.channels.trySendBlocking
